Module properties are now properly persisted between sessions
